requirements.txt
flask
psycopg2-binary==2.9.3
requests
